[Decomp][Item] Cohesion split: extract vendor/bank + enchant/socket handlers to ItemHandler*.cpp
